Add files via upload
This challenge has been adapted from The Turing School’s and Jump Start Lab’s Event Manager and updated to make use of GoogleCivic API. For the first task, I was to determine the government representatives for each attendee based on their zip code (as provided by the CSV file). For each attendee, we were then to include a customized letter that thanks them for attending and provides a list of their representatives (each attendee letter can be found in the output folder).